refactor: move `AncillaryFilesManifest` to cardano internal db
