+++ internal/cardano-node/mithril-cardano-node-internal-database/Cargo.toml
@@ -17,6 +17,7 @@ async-trait = { workspace = true }
 digest = { workspace = true }
 hex = { workspace = true }
 mithril-common = { path = "../../../mithril-common", version = "=0.5" }
+serde = { workspace = true }
 serde_json = { workspace = true }
 sha2 = "0.10.9"
 slog = { workspace = true }
