+++ internal/cardano-node/mithril-cardano-node-internal-database/src/entities/mod.rs
@@ -1,8 +1,10 @@
 //! Entities related to a Cardano node internal database.
 //!
 
+mod ancillary_files_manifest;
 mod immutable_file;
 mod ledger_state_snapshot;
 
+pub use ancillary_files_manifest::*;
 pub use immutable_file::*;
 pub use ledger_state_snapshot::*;
