+++ mithril-aggregator/src/services/snapshotter/ancillary_signer/with_gcp_kms.rs
@@ -5,8 +5,8 @@ use gcloud_kms::client::{Client as GcpKmsClient, ClientConfig};
 use gcloud_kms::grpc::kms::v1::AsymmetricSignRequest;
 use slog::{debug, Logger};
 
+use mithril_cardano_node_internal_database::entities::AncillaryFilesManifest;
 use mithril_common::crypto_helper::ManifestSignature;
-use mithril_common::entities::AncillaryFilesManifest;
 use mithril_common::StdResult;
 
 use crate::services::ancillary_signer::{AncillarySigner, GcpCryptoKeyVersionResourceName};
