+++ mithril-aggregator/src/services/snapshotter/ancillary_signer/with_secret_key.rs
@@ -1,8 +1,8 @@
 use async_trait::async_trait;
 use slog::{debug, Logger};
 
+use mithril_cardano_node_internal_database::entities::AncillaryFilesManifest;
 use mithril_common::crypto_helper::{ManifestSignature, ManifestSigner};
-use mithril_common::entities::AncillaryFilesManifest;
 use mithril_common::logging::LoggerExtensions;
 use mithril_common::StdResult;
